refactor: extract MenuBarStatusRecoveryAction.swift from MenuBarStatusPresentation.swift
Splits the `MenuBarStatusRecoveryAction` enum out of
MenuBarStatusPresentation.swift so that file holds only the
presentation struct.

Byte-identical move of lines 3-11. New file imports Foundation
(no CoreGraphics needed for the enum); MenuBarStatusPresentation.swift
keeps CoreGraphics for CGFloat.

MenuBarStatusPresentation.swift: 52 → 42 lines. Tests: 667.
