+++ Sources/BugNarrator/Utilities/MenuBarStatusPresentation.swift
@@ -1,15 +1,5 @@
 import CoreGraphics
 
-enum MenuBarStatusRecoveryAction: Equatable {
-    case none
-    case microphone
-    case screenRecording
-    case systemAudio
-    case providerSettings
-    case exportConfiguration
-    case storage
-}
-
 struct MenuBarStatusPresentation: Equatable {
     let preferredWidth: CGFloat
     let recoveryAction: MenuBarStatusRecoveryAction
