+++ Sources/BugNarrator/Utilities/MenuBarStatusRecoveryAction.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum MenuBarStatusRecoveryAction: Equatable {
+    case none
+    case microphone
+    case screenRecording
+    case systemAudio
+    case providerSettings
+    case exportConfiguration
+    case storage
+}
